Make sure trampoline size is computed once

+++ src/koffi/src/call.hh
@@ -157,6 +157,6 @@ inline T *CallData::AllocHeap(Size size, Size align)
 
 void PerformAsyncRelay(napi_env env, napi_value callback, void *ctx, void *udata);
 
-void *GetTrampoline(int16_t idx);
+void *GetTrampoline(int idx);
 
 }
